Former-commit-id: 138aeb36d225a2b94625f81188db166c23a0d3f9
Former-commit-id: 48e3413cfb15e9986f9a02880514a3f1b5fdad4d

+++ backend/src/main/java/com/moi/anitime/model/repo/AnimalRepo.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AnimalRepo extends JpaRepository<Animal, Integer> {
+public interface AnimalRepo extends JpaRepository<Animal, Long> {
 
     @Query(name = "getAnimal", nativeQuery = true)
     public List<AnimalPreviewRes> getAnimal(int generalNo, String kind, char sexcd, String sortQuery, Pageable curPageNo);
